test: extend text DropTest and move into it's own package
