+++ library/src/test/java/com/alexfu/sqlitequerybuilder/testtext/drop/DropViewTest.java
@@ -1,0 +1,33 @@
+package com.alexfu.sqlitequerybuilder.testtext.drop;
+
+import com.alexfu.sqlitequerybuilder.api.SQLiteQueryBuilder;
+import org.junit.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public final class DropViewTest {
+  @Test
+  public final void testDropView() {
+    // Arrange and Act
+    String statement = SQLiteQueryBuilder
+      .drop()
+      .view("myView")
+      .ifExists()
+      .toString();
+
+    // Assert
+    assertThat(statement).isEqualTo("DROP VIEW IF EXISTS myView");
+  }
+
+  @Test
+  public final void testDropViewNoIfExists() {
+    // Arrange and Act
+    String statement = SQLiteQueryBuilder
+      .drop()
+      .view("myView")
+      .toString();
+
+    // Assert
+    assertThat(statement).isEqualTo("DROP VIEW myView");
+  }
+}
